requirements.txt
natsort==8.4.0
pillow==11.1.0
